refactor(agent-service): centralize types and isolate clients
- Move all type/interface definitions under types/ (new types/metadata.ts,
  types/api.ts); drop duplicate LogicalPlan/PortSchema/InputPortInfo and the
  identical AgentSettingsApi/UpdateAgentSettingsRequest.
- Make api/ pure HTTP clients: extract the execution client
  (api/execution-client.ts), move JWT helpers to auth/jwt.ts and endpoint
  config to config/endpoints.ts, rename *-api.ts -> *-client.ts
  (backend-api -> operator-metadata-client).
- Rename misleading *Config types: AgentDelegateConfig -> AgentDelegation
  (+ AgentDelegationDto), TexeraAgentConfig -> TexeraAgentOptions,
  ExecutionConfig -> ExecutionRequestParams, getBackendConfig ->
  getServiceEndpoints.
- De-duplicate the table/IO-shape formatters into tools-utility.ts.
- Align CommentBox, WorkflowSettings.executionMode, and WorkflowFatalError
  with the frontend wire contract.

Refs #5563
