+++ agent-service/src/agent/util/workflow-system-metadata.ts
@@ -18,24 +18,14 @@
  */
 
 import Ajv from "ajv";
-import { fetchOperatorMetadata, type OperatorSchema, type OperatorMetadata } from "../../api/backend-api";
+import { fetchOperatorMetadata } from "../../api/operator-metadata-client";
+import type { OperatorMetadata, OperatorSchemaInfo, CompactOperatorSchema } from "../../types/metadata";
 import type { ValidationError, Validation } from "../../types/workflow";
 import { createLogger } from "../../logger";
 
 const log = createLogger("WorkflowSystemMetadata");
 
 export type { ValidationError, Validation } from "../../types/workflow";
-
-interface OperatorSchemaInfo {
-  properties: any;
-  required: any;
-  definitions: any;
-}
-
-interface CompactOperatorSchema {
-  properties: Record<string, any>;
-  required: string[];
-}
 
 const FILTERED_PROPERTY_KEYS = ["dummyPropertyList"];
