+++ agent-service/src/auth/jwt.ts
