+++ agent-service/src/index.ts
@@ -22,5 +22,5 @@ export { WorkflowState } from "./agent/workflow-state";
 export { WorkflowResultState } from "./agent/workflow-result-state";
 export { WorkflowSystemMetadata } from "./agent/util/workflow-system-metadata";
 export * from "./agent/tools";
-export { TexeraAgent, type TexeraAgentConfig, type AgentMessageResult } from "./agent/texera-agent";
+export { TexeraAgent } from "./agent/texera-agent";
 export { buildSystemPrompt } from "./agent/prompts";
