+++ agent-service/src/types/index.ts
@@ -19,4 +19,6 @@
 
 export * from "./workflow";
 export * from "./execution";
+export * from "./metadata";
 export * from "./agent";
+export * from "./api";
